added everything so far.
run the program through the main() method in flightdb_app.py--ensure that the host/port are both correct in the config. make sure to import the database first with flight_db_dump_final in MySQL workbench; the other sql files are just for context/to show how everything was done.